New job search

+++ src/Merit/Merit.Web/Pages/EditCompanyMerits.cshtml.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Merit.AdvertisementService;
 using Merit.Data.Models;
 using Merit.MeritService;
 using Microsoft.AspNetCore.Identity;
@@ -13,6 +14,7 @@ namespace Merit.Web.Pages
     public class EditCompanyMeritsModel : PageModel
     {
         private IMeritService meritService = new MeritService.MeritService();
+        private IAdvertisementService advertisementService = new AdvertisementService.AdvertisementService();
 
         private readonly UserManager<IdentityUser> userManager;
         private readonly SignInManager<IdentityUser> signInManager;
@@ -23,14 +25,14 @@ namespace Merit.Web.Pages
             this.userManager = userManager;
         }
 
-        [BindProperty(SupportsGet = true)]
+        [BindProperty]
         public List<CompanyMerit> CompanyMeritList { get; set; }
         [BindProperty]
         public CompanyMerit CMerit { get; set; }
-        
+
         [BindProperty(SupportsGet = true)]
         public int SelectedMeritID { get; set; }
-        [BindProperty]         
+        [BindProperty]
         public string CategoryText { get; set; }
         [BindProperty]
         public string SubCategoryText { get; set; }
@@ -39,8 +41,11 @@ namespace Merit.Web.Pages
 
         public string Message { get; set; }
 
-        [BindProperty(SupportsGet = true)]
+        [BindProperty]
         public bool Visi { get; set; } = false;
+
+        public bool Success { get; set; }
+
 
         public async Task<IActionResult> OnGetAsync()
         {
@@ -55,6 +60,10 @@ namespace Merit.Web.Pages
             if (cUser is CompanyUser companyUser)
             {
                 CompanyMeritList = meritService.ReadCompanyMerits(companyUser.CompanyUserId);
+                if (SelectedMeritID != 0 && CompanyMeritList.FirstOrDefault(x => x.CompanyMeritId == SelectedMeritID) == null)
+                {
+                    return Redirect($"/CompanyInfoPage");
+                }
             }
             else if (cUser is PersonalUser)
             {
@@ -70,22 +79,42 @@ namespace Merit.Web.Pages
                     DescriptionText = merit.Description;
                 }
             }
+            if (SelectedMeritID != 0 && advertisementService.MeritExistsInAdvertisement(SelectedMeritID))
+            {
+                Success = false;
+                Visi = true;
+                Message = "Obs! Denna merit är länkad till en eller flera annonser.";
+            }
             return Page();
         }
         public async Task OnPostEdit()
         {
             meritService.EditCompanyMerit(CMerit);
             Visi = true;
+            Success = true;
             Message = "Merit ändrad!";
             SelectedMeritID = 0;
             await OnGetAsync();
         }
         public async Task OnPostDelete()
         {
-            meritService.DeleteCompanyMerit(CMerit);
-            Visi = true;
-            Message = "Merit borttagen!";
-            SelectedMeritID = 0;
+
+
+            if (advertisementService.MeritExistsInAdvertisement(CMerit.CompanyMeritId))
+            {
+                Message = "Du kan inte ta bort meriten, den används i en annons.";
+                Visi = true;
+                Success = false;
+                SelectedMeritID = 0;
+            }
+            else
+            {
+                meritService.DeleteCompanyMerit(CMerit);
+                Visi = true;
+                Success = true;
+                Message = "Merit borttagen!";
+                SelectedMeritID = 0;
+            }
             await OnGetAsync();
         }
     }
